improved formatting and spacing for activities and began formatting summary section.
