+++ client/src/components/Activity/CalendarTile.tsx
@@ -1,0 +1,5 @@
+import Calendar from "react-calendar";
+
+export default function MyCalendar() {
+  return <Calendar calendarType="gregory" />;
+}
